♻️🚚: move vsphere test into package
