+++ pkg/vsphere/vsphere_test.go
@@ -1,0 +1,13 @@
+package vsphere
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestSpecRunner(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "vsphere tests")
+}
